Skip copying a texture if the file exists.
If a file with the same name exists in the export directory, skip
copying it rather than attempting to copy and throwing the error:

```
shutil.SameFileError: '/tmp/example/texture.png' and
'/tmp/example/texture.png' are the same file
```

This error is common if you share a texture among multiple models.

Fixes #182.